requirements.txt
pandas
faker
requests
python-dotenv
sqlalchemy
psycopg2-binary